requirements.txt
numpy
pillow
opencv-python
pathlib
albumentations
pycocotools